Correct spectacle config
- add to install script

[#155302342]

+++ setup-spectacle.sh
@@ -5,4 +5,6 @@ preferences_path="/Users/$(whoami)/Library/Preferences/com.divisiblebyzero.Spect
 echo $current_directory
 echo $preferences_path
 
-ln -s "${current_directory}/assets/com.divisiblebyzero.Spectacle.plist" "${preferences_path}"
+if [[ ! -e "${preferences_path}" ]]; then
+	ln -s "${current_directory}/assets/com.divisiblebyzero.Spectacle.plist" "${preferences_path}"
+fi
